quartz/requirements.txt
aerospike>=10.0.0
fabric>=3.0.0
pytest>=7.2.1
pytest-html>=3.2.0
pytest-logger>=0.5.1
pytest-xdist>=3.1.0
requests>=2.28.2
boto3~=1.26.78
botocore~=1.29.78
mysql-connector-python>=1.2
redis>=5.0.1
pylint>=3.0.3